requirements.txt
numpy
googlemaps
pandas
python-dotenv
folium
tkinterweb
pillow
pytest
pyinstaller
setuptools